moving character-sheet components that were in the wrong folder, also increase the size of the feature box
